+++ spec/puppet-check/ruby_parser_spec.rb
@@ -17,22 +17,12 @@ describe RubyParser do
       expect(PuppetCheck.warning_files).to eql([])
       expect(PuppetCheck.clean_files).to eql([])
     end
-    if Gem::Version.new(RUBY_VERSION.dup) >= Gem::Version.new('2.1.0')
-      it 'puts a bad style ruby file in the warning files array' do
-        PuppetCheck.style_check = true
-        RubyParser.ruby([fixtures_dir + 'lib/style.rb'])
-        expect(PuppetCheck.error_files).to eql([])
-        expect(PuppetCheck.warning_files[0]).to match(%r{^#{fixtures_dir}lib/style.rb:\n.*Useless assignment.*\n.*Use the new.*\n.*Do not introduce.*\n.*Prefer single.*\n.*is a writable attribute.*\n.*Issue has no descriptive comment})
-        expect(PuppetCheck.clean_files).to eql([])
-      end
-    else
-      it 'puts a bad style ruby file in the warning files array' do
-        PuppetCheck.style_check = true
-        RubyParser.ruby([fixtures_dir + 'lib/style.rb'])
-        expect(PuppetCheck.error_files).to eql([])
-        expect(PuppetCheck.warning_files[0]).to match(%r{^#{fixtures_dir}lib/style.rb:\n.*Useless assignment.*\n.*Use the new.*\n.*Do not introduce.*\n.*Prefer single})
-        expect(PuppetCheck.clean_files).to eql([])
-      end
+    it 'puts a bad style ruby file in the warning files array' do
+      PuppetCheck.style_check = true
+      RubyParser.ruby([fixtures_dir + 'lib/style.rb'])
+      expect(PuppetCheck.error_files).to eql([])
+      expect(PuppetCheck.warning_files[0]).to match(%r{^#{fixtures_dir}lib/style.rb:\n.*Useless assignment.*\n.*Use the new.*\n.*Do not introduce.*\n.*Prefer single.*\n.*is a writable attribute.*\n.*Issue has no descriptive comment})
+      expect(PuppetCheck.clean_files).to eql([])
     end
     it 'puts a bad style ruby file in the clean files array when rubocop_args ignores its warnings' do
       PuppetCheck.style_check = true