+++ spec/puppet-check/ruby_parser_spec.rb
@@ -6,35 +6,31 @@ describe RubyParser do
     PuppetCheck.error_files = []
     PuppetCheck.warning_files = []
     PuppetCheck.clean_files = []
-    PuppetCheck.style_check = false
     PuppetCheck.rubocop_args = []
   end
 
   context '.ruby' do
     it 'puts a bad syntax ruby file in the error files array' do
-      RubyParser.ruby([fixtures_dir + 'lib/syntax.rb'])
+      RubyParser.ruby([fixtures_dir + 'lib/syntax.rb'], false)
       expect(PuppetCheck.error_files[0]).to match(%r{^#{fixtures_dir}lib/syntax.rb:\n.*syntax error})
       expect(PuppetCheck.warning_files).to eql([])
       expect(PuppetCheck.clean_files).to eql([])
     end
     it 'puts a bad style ruby file in the warning files array' do
-      PuppetCheck.style_check = true
-      RubyParser.ruby([fixtures_dir + 'lib/style.rb'])
+      RubyParser.ruby([fixtures_dir + 'lib/style.rb'], true)
       expect(PuppetCheck.error_files).to eql([])
       expect(PuppetCheck.warning_files[0]).to match(%r{^#{fixtures_dir}lib/style.rb:\n.*Useless assignment.*\n.*Use the new.*\n.*Do not introduce.*\n.*Prefer single.*\n.*is a writable attribute.*\n.*Issue has no descriptive comment})
       expect(PuppetCheck.clean_files).to eql([])
     end
     it 'puts a bad style ruby file in the clean files array when rubocop_args ignores its warnings' do
-      PuppetCheck.style_check = true
       PuppetCheck.rubocop_args = ['--except', 'Lint/UselessAssignment,Style/HashSyntax,Style/GlobalVars,Style/StringLiterals']
-      RubyParser.ruby([fixtures_dir + 'lib/rubocop_style.rb'])
+      RubyParser.ruby([fixtures_dir + 'lib/rubocop_style.rb'], true)
       expect(PuppetCheck.error_files).to eql([])
       expect(PuppetCheck.warning_files).to eql([])
       expect(PuppetCheck.clean_files).to eql(["#{fixtures_dir}lib/rubocop_style.rb"])
     end
     it 'puts a good ruby file in the clean files array' do
-      PuppetCheck.style_check = true
-      RubyParser.ruby([fixtures_dir + 'lib/good.rb'])
+      RubyParser.ruby([fixtures_dir + 'lib/good.rb'], true)
       expect(PuppetCheck.error_files).to eql([])
       expect(PuppetCheck.warning_files).to eql([])
       expect(PuppetCheck.clean_files).to eql(["#{fixtures_dir}lib/good.rb"])
@@ -70,29 +66,26 @@ describe RubyParser do
 
   context '.librarian' do
     it 'puts a bad syntax librarian Puppet file in the error files array' do
-      RubyParser.librarian([fixtures_dir + 'librarian_syntax/Puppetfile'])
+      RubyParser.librarian([fixtures_dir + 'librarian_syntax/Puppetfile'], false)
       expect(PuppetCheck.error_files[0]).to match(%r{^#{fixtures_dir}librarian_syntax/Puppetfile:\n.*syntax error})
       expect(PuppetCheck.warning_files).to eql([])
       expect(PuppetCheck.clean_files).to eql([])
     end
     it 'puts a bad style librarian Puppet file in the warning files array' do
-      PuppetCheck.style_check = true
-      RubyParser.librarian([fixtures_dir + 'librarian_style/Puppetfile'])
+      RubyParser.librarian([fixtures_dir + 'librarian_style/Puppetfile'], true)
       expect(PuppetCheck.error_files).to eql([])
       expect(PuppetCheck.warning_files[0]).to match(%r{^#{fixtures_dir}librarian_style/Puppetfile:\n.*Align the parameters.*\n.*Use the new})
       expect(PuppetCheck.clean_files).to eql([])
     end
     it 'puts a bad style librarian Puppet file in the clean files array when rubocop_args ignores its warnings' do
-      PuppetCheck.style_check = true
       PuppetCheck.rubocop_args = ['--except', 'Style/AlignParameters,Style/HashSyntax']
-      RubyParser.librarian([fixtures_dir + 'librarian_style/Puppetfile'])
+      RubyParser.librarian([fixtures_dir + 'librarian_style/Puppetfile'], true)
       expect(PuppetCheck.error_files).to eql([])
       expect(PuppetCheck.warning_files).to eql([])
       expect(PuppetCheck.clean_files).to eql(["#{fixtures_dir}librarian_style/Puppetfile"])
     end
     it 'puts a good librarian Puppet file in the clean files array' do
-      PuppetCheck.style_check = true
-      RubyParser.librarian([fixtures_dir + 'librarian_good/Puppetfile'])
+      RubyParser.librarian([fixtures_dir + 'librarian_good/Puppetfile'], true)
       expect(PuppetCheck.error_files).to eql([])
       expect(PuppetCheck.warning_files).to eql([])
       expect(PuppetCheck.clean_files).to eql(["#{fixtures_dir}librarian_good/Puppetfile"])